test: logging-gate compliance + tools test consolidation
Strip server-side PostHog capture from the Vercel redirect endpoints
(docs/logging.md pins the public site to browser-only capture; the
install dialog's client events are the funnel signal), document
ade_tool_fetched and the install-dialog taxonomy, add ade_tool_fetched
to the dashboard event catalog, and fold the lock + retry test files
into install.test.ts per the per-folder test budget.
